+++ src/pages/Main.tsx
@@ -2,12 +2,11 @@ import React from "react";
 import { Routes, Route } from "react-router-dom";
 
 import ErrorPage from "../common/components/ErrorPage";
-import BasicPageLayout from "../common/components/layouts/BasicPageLayout";
 
 import useUserStore from "../store/user.zustand";
 import Routine from "./routine/Routine";
 import BottomNav from "./BottomNav";
-import Profile from "./profile/Profile";
+import ProfileRoute from "./profile/ProfileRoute";
 
 const Main = () => {
   let username = useUserStore((state) => state.username);
@@ -21,7 +20,7 @@ const Main = () => {
         <Route path="group" element={<div>group</div>} />
         <Route path="routine/*" element={<Routine />} />
         <Route path="statistics" element={<div>statistics</div>} />
-        <Route path="profile/:username" element={<Profile />} />
+        <Route path="profile/*" element={<ProfileRoute />} />
         <Route path="*" element={<ErrorPage />} />
       </Routes>
       <BottomNav username={username} />
